splitting popop and use props for display popup
